Refactor integration tests for DynamoDB backend to use base tests
